change to fix Sun compiler
git-svn-id: http://svn.boost.org/svn/boost/trunk@20461 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/libs/date_time/src/gregorian/greg_month.cpp
@@ -31,7 +31,7 @@ namespace gregorian {
 
     if(month_map_ptr->empty()) {
       std::string s("");
-      for(int i = 1; i <= 12; ++i) {
+      for(unsigned short i = 1; i <= 12; ++i) {
         greg_month m(static_cast<month_enum>(i));
         s = m.as_long_string();
 #if defined(BOOST_DATE_TIME_NO_STD_TRANSFORM)
